Created SettingsHistory activity.
