+++ app/src/main/java/me/ccrama/redditslide/Activities/SettingsGeneral.java
@@ -165,20 +165,6 @@ public class SettingsGeneral extends BaseActivityAnim {
                 }
             });
         }
-
-        {
-            SwitchCompat single = (SwitchCompat) findViewById(R.id.scrollseen);
-
-            single.setChecked(SettingValues.scrollSeen);
-            single.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-                @Override
-                public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                    SettingValues.scrollSeen = isChecked;
-                    SettingValues.prefs.edit().putBoolean(SettingValues.PREF_SCROLL_SEEN, isChecked).apply();
-
-                }
-            });
-        }
         {
             SwitchCompat single = (SwitchCompat) findViewById(R.id.datause);
 
@@ -335,50 +321,6 @@ public class SettingsGeneral extends BaseActivityAnim {
                 }
             });
         }
-
-        SwitchCompat storeHistory = ((SwitchCompat) findViewById(R.id.storehistory));
-        storeHistory.setChecked(SettingValues.storeHistory);
-        storeHistory.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                SettingValues.storeHistory = isChecked;
-                SettingValues.prefs.edit().putBoolean(SettingValues.PREF_STORE_HISTORY, isChecked).apply();
-
-                new AlertDialog.Builder(SettingsGeneral.this)
-                        .setTitle("Delete entry")
-                        .setMessage("storeHistory is now: " + Boolean.toString(SettingValues.storeHistory))
-                        .setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int which) {
-                                // continue with delete
-                            }
-                        })
-                        .setNegativeButton(android.R.string.no, new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int which) {
-                                // do nothing
-                            }
-                        })
-                        .setIcon(android.R.drawable.ic_dialog_alert)
-                        .show();
-
-                if (isChecked) {
-                    findViewById(R.id.scrollseen).setEnabled(true);
-                    findViewById(R.id.storensfw).setEnabled(true);
-                }
-                else {
-                    ((SwitchCompat) findViewById(R.id.storensfw)).setChecked(false);
-                    ((SwitchCompat) findViewById(R.id.storensfw)).setEnabled(false);
-                    SettingValues.storeNSFWHistory = false;
-                    SettingValues.prefs.edit().putBoolean(SettingValues.PREF_STORE_NSFW_HISTORY, false).apply();
-
-                    ((SwitchCompat) findViewById(R.id.scrollseen)).setChecked(false);
-                    ((SwitchCompat) findViewById(R.id.scrollseen)).setEnabled(false);
-                    SettingValues.scrollSeen = false;
-                    SettingValues.prefs.edit().putBoolean(SettingValues.PREF_SCROLL_SEEN, false).apply();
-                }
-            }
-        });
-
-        ((SwitchCompat)findViewById(R.id.storensfw)).setChecked(SettingValues.storeNSFWHistory);
     }
 
 }
