+++ app/src/main/res/layout/activity_settings.xml
@@ -75,6 +75,30 @@
                 </LinearLayout>
 
                 <LinearLayout
+                        android:id="@+id/history"
+                        android:layout_width="match_parent"
+                        android:layout_height="56dp"
+                        android:background="?android:selectableItemBackground"
+                        android:orientation="horizontal">
+
+                    <ImageView
+                            android:layout_width="56dp"
+                            android:layout_height="match_parent"
+                            android:layout_marginRight="16dp"
+                            android:padding="16dp"
+                            app:srcCompat="@drawable/edit"
+                            android:tint="?attr/tint" />
+
+                    <TextView
+                            android:layout_width="wrap_content"
+                            android:layout_height="match_parent"
+                            android:gravity="center_vertical"
+                            android:text="@string/settings_title_history"
+                            android:textColor="?attr/font"
+                            android:textSize="16sp" />
+                </LinearLayout>
+
+                <LinearLayout
                     android:id="@+id/handling"
                     android:layout_width="match_parent"
                     android:layout_height="56dp"
