+++ app/src/main/res/layout/activity_settings_general.xml
@@ -210,128 +210,6 @@
             </RelativeLayout>
 
             <RelativeLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="56dp"
-                    android:gravity="center_vertical"
-                    android:orientation="horizontal"
-                    android:paddingEnd="16dp"
-                    android:background="?android:selectableItemBackground"
-                    android:paddingStart="16dp">
-
-                <LinearLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_centerVertical="true"
-                        android:orientation="vertical">
-
-                    <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:text="@string/settings_store_history"
-                            android:textColor="?attr/font"
-                            android:textSize="16sp" />
-
-                </LinearLayout>
-
-                <android.support.v7.widget.SwitchCompat
-                        android:id="@+id/storehistory"
-                        android:layout_width="match_parent"
-
-                        android:layout_height="match_parent"
-                        android:background="?android:selectableItemBackground"
-                        android:backgroundTint="?attr/tint"
-                        android:button="@null"
-                        android:buttonTint="?attr/tint"
-                        android:hapticFeedbackEnabled="true"
-                        android:textColor="?attr/font"
-                        android:textColorHint="?attr/font" android:checked="true"/>
-
-            </RelativeLayout>
-
-            <RelativeLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="56dp"
-                    android:gravity="center_vertical"
-                    android:orientation="horizontal"
-                    android:paddingEnd="16dp"
-                    android:background="?android:selectableItemBackground"
-                    android:paddingStart="16dp">
-
-                <LinearLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_centerVertical="true"
-                        android:orientation="vertical">
-
-                    <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:text="@string/settings_store_nsfw_history"
-                            android:textColor="?attr/font"
-                            android:textSize="16sp" />
-
-                </LinearLayout>
-
-                <android.support.v7.widget.SwitchCompat
-                        android:id="@+id/storensfw"
-                        android:layout_width="match_parent"
-
-                        android:layout_height="match_parent"
-                        android:background="?android:selectableItemBackground"
-                        android:backgroundTint="?attr/tint"
-                        android:button="@null"
-                        android:buttonTint="?attr/tint"
-                        android:hapticFeedbackEnabled="true"
-                        android:textColor="?attr/font"
-                        android:textColorHint="?attr/font" />
-
-            </RelativeLayout>
-
-            <RelativeLayout
-                android:layout_width="match_parent"
-                android:layout_height="56dp"
-                android:gravity="center_vertical"
-                android:orientation="horizontal"
-                android:paddingEnd="16dp"
-                android:background="?android:selectableItemBackground"
-                android:paddingStart="16dp">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_centerVertical="true"
-                    android:orientation="vertical">
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="@string/settings_general_scroll_read"
-                        android:textColor="?attr/font"
-                        android:textSize="16sp" />
-
-                </LinearLayout>
-
-                <android.support.v7.widget.SwitchCompat
-                    android:id="@+id/scrollseen"
-                    android:layout_width="match_parent"
-
-                    android:layout_height="match_parent"
-                    android:background="?android:selectableItemBackground"
-                    android:backgroundTint="?attr/tint"
-                    android:button="@null"
-                    android:buttonTint="?attr/tint"
-                    android:hapticFeedbackEnabled="true"
-                    android:textColor="?attr/font"
-                    android:textColorHint="?attr/font" />
-
-            </RelativeLayout>
-            <View
-                android:layout_width="match_parent"
-                android:background="?attr/tint"
-                android:alpha=".25"
-                android:layout_height="0.25dp"/>
-
-            <RelativeLayout
                 android:layout_width="match_parent"
                 android:layout_height="56dp"
                 android:gravity="center_vertical"
