+++ app/src/main/res/values/donottranslate.xml
@@ -1,5 +1,4 @@
 <resources>
-    <string name="hello_world">Hello world!</string>
 
     <!-- Text for the example submissions
          Used in the edit card view screen  -->
